Don't include .vscode/c_cpp_properties.json in the repository
When the CMake Tools extension is used with the C/C++ extension,
this file ends up with user-specific absolute path names, which
are filled in by the tools.  So it's essentially an auto-generated
file, and as such should not be included in the repository.

When you first load the solution into VSCode, you may be faced with
one or more prompts, and responding to these may generate
the c_cpp_properties.json file for you.  You may want to customize
it to match your personal preferences.