Fork zlib to keep working tree clean
A version of zconf.h is checked into the upstream repository that
interferes with the zconf.h auto-generated when building with
build systems generated with CMake.  To avoid the conflict, the build
system will rename this zconf.h, resulting in the working tree having
one deleted file and one untracked file.

This fork of the repository has the zconf.h removed to avoid
the issue entirely.